Make passing result more flexible
Allow an operation to pass it’s result to multiple others, and allow an operation to take multiple inputs.
